game 데이터 등록기능 구현 시작 (이진규)

+++ src/main/java/org/zerock/guestbook/entity/BoardGame.java
@@ -21,7 +21,7 @@ public class BoardGame {
     private String writer;
 
     @Column(name = "B_G_WRITERNUM", nullable = false)
-    private Integer writerNum;
+    private Long writerNum;
 
     @Column(name = "B_G_TITLE", nullable = false)
     private String title;
